Rebbit reconfigure on manager.
