+++ .env
@@ -1,6 +1,6 @@
 WORKER_BASE_URL=http://crackhash-worker
 MANAGER_URL=http://crackhash-manager
-WORKERS_COUNT=1
+WORKERS_COUNT=2
 TIMEOUT=3
 
 RABBIT_USER=user
