+++ manager/src/main/kotlin/ru/nsu/fit/crackhash/manager/controller/ManagerInternalController.kt
@@ -1,15 +1,12 @@
 package ru.nsu.fit.crackhash.manager.controller
 
-import org.springframework.web.bind.annotation.PatchMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.amqp.rabbit.annotation.RabbitListener
+import org.springframework.stereotype.Component
 import ru.nsu.fit.crackhash.manager.model.dto.WorkerResponseDto
 import ru.nsu.fit.crackhash.manager.service.ManagerInternalService
 
-@RestController
-@RequestMapping("/internal/api/manager/hash/crack/request")
+@Component
 class ManagerInternalController(private val managerInternalService: ManagerInternalService) {
-    @PatchMapping
-    fun crackRequest(@RequestBody response: WorkerResponseDto) = managerInternalService.crackRequest(response)
+    @RabbitListener(queues = ["worker-to-manager"])
+    fun crackRequest(response: WorkerResponseDto) = managerInternalService.crackRequest(response)
 }
