+++ manager/src/main/kotlin/ru/nsu/fit/crackhash/manager/service/WorkerService.kt
@@ -1,0 +1,5 @@
+package ru.nsu.fit.crackhash.manager.service
+
+import ru.nsu.fit.crackhash.manager.model.entity.WorkerEntity
+
+interface WorkerService: Iterable<WorkerEntity>
