+++ manager/src/main/kotlin/ru/nsu/fit/crackhash/manager/service/impl/SendServiceImpl.kt
@@ -10,12 +10,13 @@ import ru.nsu.fit.crackhash.manager.model.dto.WorkerTaskDto
 import ru.nsu.fit.crackhash.manager.model.entity.WorkerEntity
 import ru.nsu.fit.crackhash.manager.repo.TaskRepo
 import ru.nsu.fit.crackhash.manager.service.SendService
+import ru.nsu.fit.crackhash.manager.service.WorkerService
 
 @Service
 class SendServiceImpl(
     @Value("\${workers.count}")
     private val partCount: Int,
-    private val workers: List<WorkerEntity>,
+    private val workers: WorkerService,
     private val taskRepo: TaskRepo,
     private val logger: Logger,
 ) : SendService {
