+++ worker/src/main/kotlin/ru/nsu/fit/crackhash/worker/service/impl/TaskExecutorServiceImpl.kt
@@ -35,7 +35,7 @@ class TaskExecutorServiceImpl(
 
             manager.convertAndSend(
                 exchange.name,
-                "worker-to-manager",
+                "manager",
                 WorkerResponseDto(
                     workerTask.partNumber,
                     workerTask.requestId,
