+++ worker/src/main/resources/application.yml
@@ -2,6 +2,8 @@ manager:
   url: ${MANAGER_URL}
   timeout: ${TIMEOUT}
 
+worker.number: ${WORKER_NUMBER}
+
 spring:
   rabbitmq:
     host: rabbitmq
